refactor tests
Related to: #319
